Move addXRPackageScript to _withPage
