+++ tests/utils/_withPage.js
@@ -1,6 +1,15 @@
 const puppeteer = require('puppeteer');
 
-const { addXRPackageScript } = require('./_testHelpers');
+const addXRPackageScript = async (page, port) => {
+  await page.addScriptTag({
+    type: 'module',
+    content: `
+      import { XRPackage, XRPackageEngine } from "http://localhost:${port}/xrpackage.js";
+      window.XRPackage = XRPackage;
+      window.XRPackageEngine = XRPackageEngine;
+    `,
+  });
+};
 
 module.exports = async (t, run) => {
   const browser = await puppeteer.launch({
